refactor compile tests
